+++ ucw/io.h
@@ -65,13 +65,13 @@ ucw_off_t ucw_file_size(const char *name);
 
 /* io-mmap.c */
 
-void *mmap_file(const char *name, unsigned *len, int writeable);
-void munmap_file(void *start, unsigned len);
+void *mmap_file(const char *name, size_t *len, int writeable);
+void munmap_file(void *start, size_t len);
 
 /* io-careful.c */
 
-int careful_read(int fd, void *buf, int len);
-int careful_write(int fd, const void *buf, int len);
+int careful_read(int fd, void *buf, size_t len);
+int careful_write(int fd, const void *buf, size_t len);
 
 /* io-sync.c */
